Failing still

+++ deno/rest_api_class/Service/Service.ts
@@ -1,39 +1,67 @@
+import { ServiceIdNotNumber } from "/Service/Service.errors.ts";
+import { IModel } from "/Models/Model.interface.ts";
 import { IRepository } from "/Repository/Repository.interface.ts";
-import { IModel } from "/Models/Model.ts";
-
-export interface IService<MODEL extends IModel> {
-  repo: IRepository<MODEL>;
-  getAll(): Promise<MODEL[]>;
-  getOne(model: Partial<MODEL>): Promise<MODEL>;
-  create(model: Partial<MODEL>): Promise<MODEL>;
-  update(model: Partial<MODEL>): Promise<MODEL>;
-  delete(model: Partial<MODEL>): Promise<MODEL>;
-}
+import { IService } from "/Service/Service.interface.ts";
 
 export abstract class Service<MODEL extends IModel> implements IService<MODEL> {
   public repo: IRepository<MODEL>;
+  public model: new () => MODEL;
 
-  constructor(repo: IRepository<MODEL>) {
+  constructor(repo: IRepository<MODEL>, model: new () => MODEL) {
     this.repo = repo;
+    this.model = model;
+  }
+
+  validateBody(body: unknown) {
+    const model = new this.model();
+    model.validate(body);
   }
 
   getAll(): Promise<MODEL[]> {
     return this.repo.getAll();
   }
 
-  getOne(model: Partial<MODEL>): Promise<MODEL> {
+  getById(id: string): Promise<MODEL> {
+    const model = new this.model();
+    const modelId = Number(id);
+
+    if (Number.isNaN(modelId)) {
+      throw new ServiceIdNotNumber();
+    }
+
+    model.id = modelId;
+
     return this.repo.getOneBy(model);
   }
 
   create(model: Partial<MODEL>): Promise<MODEL> {
+    this.validateBody(model);
     return this.repo.insertOne(model);
   }
 
-  update(model: Partial<MODEL>): Promise<MODEL> {
+  update(id: string, model: Partial<MODEL>): Promise<MODEL> {
+    const modelId = Number(id);
+
+    if (Number.isNaN(modelId)) {
+      throw new ServiceIdNotNumber();
+    }
+
+    model.id = modelId;
+
+    this.validateBody(model);
     return this.repo.updateById(model);
   }
 
-  delete(model: Partial<MODEL>): Promise<MODEL> {
+  deleteById(id: string): Promise<MODEL> {
+    const model = new this.model();
+    const modelId = Number(id);
+
+    if (Number.isNaN(modelId)) {
+      throw new ServiceIdNotNumber();
+    }
+
+    model.id = modelId;
+
     return this.repo.deleteById(model);
   }
 }
